+++ src/fastman/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 from .commands.base import Command, register
 from .console import Output